POC of a fully funtional multi agentic chatbot
This notebook now demonstrates the working of 3 expert agents being
coordinated by a special coordinator agent. The agents seamlessly work
together to response the user query which might involve accessing
company's database through sql agent as well as retrieval of related
chunks from the vector store through RAG. Finally, the consolidated
responses are fetched into the specialized customer chat support agent
who responsds to the user in  a professional and friendly tone.

There's further room of improvement in prompt emgineering as the sql
agent is slightly deviating from its role and suggesting things which it
shouldn't. So we will need to tweak the prompts for agents a little bit.

Signed-off-by: Tariq Mehmood <tariq9mehmood9@gmail.com>